+++ src/lib/ui/events.gen.ts
@@ -1,7 +1,7 @@
 import type {Gen} from '@feltcoop/gro/dist/gen/gen.js';
 import {toRootPath} from '@feltcoop/gro/dist/paths.js';
 
-import {eventInfos} from '$lib/vocab/events';
+import {eventInfos} from '$lib/vocab/event/events';
 
 // Outputs a file with services metadata that can be imported from the client.
 export const gen: Gen = async ({originId}) => {
