+++ src/lib/ui/services.ts
@@ -1,6 +1,6 @@
 import {parseServiceEventInfo} from '$lib/vocab/event/event';
 import type {ServiceEventInfo} from '$lib/vocab/event/event';
-import {eventInfoByName} from '$lib/vocab/events';
+import {eventInfoByName} from '$lib/vocab/event/events';
 
 export const findService = (name: string): ServiceEventInfo | undefined =>
 	parseServiceEventInfo(eventInfoByName.get(name));
